requirements.txt
anyio==3.6.2
asgiref==3.6.0
asyncpg==0.27.0
black==23.3.0
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
ConfigArgParse==1.5.3
Deprecated==1.2.13
dnspython==2.3.0
docstring-parser==0.15
email-validator==2.0.0.post2
fastapi==0.95.1
Flask==2.2.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==22.10.2
geventhttpclient==2.0.9
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
importlib-metadata==6.0.1
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.15.1
MarkupSafe==2.1.2
msgpack==1.0.5
mypy-extensions==1.0.0
opentelemetry-api==1.17.0
opentelemetry-instrumentation==0.38b0
opentelemetry-instrumentation-asgi==0.38b0
opentelemetry-instrumentation-asyncpg==0.38b0
opentelemetry-instrumentation-aws-lambda==0.38b0
opentelemetry-instrumentation-dbapi==0.38b0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-instrumentation-flask==0.38b0
opentelemetry-instrumentation-jinja2==0.38b0
opentelemetry-instrumentation-logging==0.38b0
opentelemetry-instrumentation-requests==0.38b0
opentelemetry-instrumentation-sqlite3==0.38b0
opentelemetry-instrumentation-system-metrics==0.38b0
opentelemetry-instrumentation-tortoiseorm==0.38b0
opentelemetry-instrumentation-urllib==0.38b0
opentelemetry-instrumentation-urllib3==0.38b0
opentelemetry-instrumentation-wsgi==0.38b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-sdk==1.17.0
opentelemetry-semantic-conventions==0.38b0
opentelemetry-util-http==0.38b0
packaging==23.1
pathspec==0.11.1
piccolo==0.109.0
piccolo-api==0.54.0
platformdirs==3.2.0
psutil==5.9.5
pydantic==1.10.7
PyJWT==2.6.0
python-multipart==0.0.6
pyzmq==25.0.2
requests==2.28.2
roundrobin==0.0.4
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
targ==0.3.8
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
